Refactored the way kNet gets included to allow a cleaner SDK generation. Few changes were made to the kNet source code.
