+++ Source/Engine/Network/CMakeLists.txt
@@ -24,4 +24,4 @@
 define_source_files (PARENT_SCOPE)
 
 # Define dependency libs
-set (ENGINE_INCLUDE_DIRS_ONLY ${ENGINE_INCLUDE_DIRS_ONLY} ../ThirdParty/kNet/include ../ThirdParty/Civetweb/include PARENT_SCOPE)
+set (ENGINE_INCLUDE_DIRS_ONLY ${ENGINE_INCLUDE_DIRS_ONLY} ../ThirdParty/kNet/include ../ThirdParty/kNet/include/kNet ../ThirdParty/Civetweb/include PARENT_SCOPE)
