+++ Source/Engine/Network/Connection.h
@@ -29,8 +29,8 @@
 #include "Timer.h"
 #include "VectorBuffer.h"
 
-#include <kNetFwd.h>
-#include <kNet/SharedPtr.h>
+#include <kNet/kNetFwd.h>
+#include <kNet/kNet/SharedPtr.h>
 
 #ifdef SendMessage
 #undef SendMessage
